Backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
SQLAlchemy==2.0.43